- Include `GlossyGradients` as part of the `core` module   - Bought the rights to this shader from the official Author (on Fiverr) -> Licensed as MIT - Improve README to better give credits
